chore: removed unused spark
