--- data-pipeline/k8s/spark-job.yaml
@@ -1,18 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: spark-etl
-  labels:
-    app: spark-etl
-spec:
-  backoffLimit: 2
-  template:
-    spec:
-      # Use the IRSA-enabled ServiceAccount
-      serviceAccountName: spark-runner
-
-      containers:
-      - name: spark-etl
-        image: ${ECR_URI}
-
-      restartPolicy: Never
