--- data-pipeline/tests/unit/test_csv_to_parquet.py
@@ -1,4 +1,0 @@
-def test_csv_to_parquet_placeholder():
-    # TODO: Implement unit tests for CSV to Parquet conversion
-    # This is a placeholder test to ensure the test suite runs
-    assert True
