--- helm/templates/serviceaccount.yaml
@@ -1,37 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: spark-runner
-  labels:
-    app: {{ .Chart.Name }}
-    chart: {{ .Chart.Name }}-{{ .Chart.Version }}
-    release: {{ .Release.Name }}
-automountServiceAccountToken: true
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: {{ .Chart.Name }}-spark-runner
-rules:
-- apiGroups: [""]
-  resources: ["pods"]
-  verbs: ["create", "get", "list", "delete", "watch"]
-- apiGroups: [""]
-  resources: ["services"]
-  verbs: ["create", "get", "list", "delete"]
-- apiGroups: ["extensions", "apps"]
-  resources: ["deployments"]
-  verbs: ["create", "get", "list", "delete"]
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: {{ .Chart.Name }}-spark-runner
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: {{ .Chart.Name }}-spark-runner
-subjects:
-- kind: ServiceAccount
-  name: spark-runner
-  namespace: {{ .Release.Namespace }}
